refactor: put IPS into separate module
